CPM-1016: Create conditions contexts
