+++ components/identifier-generator/back/tests/Acceptance/Context/CreateIdentifierGeneratorContext.php
@@ -218,77 +218,6 @@ final class CreateIdentifierGeneratorContext implements Context
         $this->tryToCreateGenerator(conditions: \array_fill(0, \intval($count), $this->getValidCondition('simple_select')));
     }
 
-    /**
-     * @When I try to create an identifier generator with 2 enabled conditions
-     */
-    public function iTryToCreateAnIdentifierGeneratorWith2EnabledConditions(): void
-    {
-        $this->tryToCreateGenerator(conditions: [
-            ['type' => 'enabled', 'value' => true],
-            ['type' => 'enabled', 'value' => true],
-        ]);
-    }
-
-    /**
-     * @When I try to create an identifier generator with 2 family conditions
-     */
-    public function iTryToCreateAnIdentifierGeneratorWith2FamilyConditions(): void
-    {
-        $this->tryToCreateGenerator(conditions: [
-            ['type' => 'family', 'operator' => 'EMPTY'],
-            ['type' => 'family', 'operator' => 'NOT EMPTY'],
-        ]);
-    }
-
-    /**
-     * @When /^I try to create an identifier generator \
-     *     with an? (?P<type>simple_select|multi_select|family|enabled) condition\
-     *     (?:(?: with| and|,) (?P<attributeCode>[^ ]*) attribute)?\
-     *     (?:(?: with| and|,) (?P<operator>[^ ]*) operator)?\
-     *     (?:(?: with| and|,) (?P<scope>[^ ]*) scope)?\
-     *     (?:(?: with| and|,) (?P<locale>[^ ]*) locale)?\
-     *     (?:(?: with| and|,) (?P<value>.*) as value)?\
-     *     (?P<unknown>(?: with| and|,) an unknown property)?$/
-     */
-    public function iTryToCreateAnIdentifierGeneratorWithCondition(
-        string $type,
-        string $attributeCode = '',
-        string $operator = '',
-        string $scope = '',
-        string $locale = '',
-        string $value = '',
-        string $unknown = '',
-    ): void {
-        $defaultCondition = $this->getValidCondition($type);
-        if ($attributeCode !== '') {
-            $defaultCondition['attributeCode'] = $attributeCode;
-        }
-        if ('undefined' === $scope) {
-            unset($defaultCondition['scope']);
-        } elseif ('' !== $scope) {
-            $defaultCondition['scope'] = $scope;
-        }
-        if ('undefined' === $locale) {
-            unset($defaultCondition['locale']);
-        } elseif ('' !== $locale) {
-            $defaultCondition['locale'] = $locale;
-        }
-        if ('undefined' === $value) {
-            unset($defaultCondition['value']);
-        } elseif ($value !== '') {
-            $defaultCondition['value'] = \json_decode($value);
-        }
-        if ('undefined' === $operator) {
-            unset($defaultCondition['operator']);
-        } elseif ($operator !== '') {
-            $defaultCondition['operator'] = $operator;
-        }
-        if ($unknown !== '') {
-            $defaultCondition['unknown'] = 'unknown property';
-        }
-        $this->tryToCreateGenerator(conditions: [$defaultCondition]);
-    }
-
     private function tryToCreateGenerator(
         ?string $code = null,
         ?array $structure = null,
